@@ -14,7 +14,7 @@ android {
             storePassword 'angcyo'
         }
     }
-    compileSdkVersion 27
+    compileSdkVersion 28
     //buildToolsVersion "23.0.3"
     defaultConfig {
         applicationId "com.angcyo.y2tools"
@@ -43,9 +43,11 @@ android {
     }
 }
 
+apply from: 'https://gitee.com/angcyo/gradle/raw/master/common.gradle'
 apply from: 'https://gitee.com/angcyo/gradle/raw/master/apk.gradle'
 
 dependencies {
-    implementation 'com.android.support:appcompat-v7:23.+'
+    //implementation 'com.android.support:appcompat-v7:23.+'
+    implementation "androidx.appcompat:appcompat:1.3.1"
     implementation files('libs/locSDK_4.2.jar')
 }
