@@ -1,6 +1,7 @@
 package com.angcyo.base;
 
-import android.support.v4.app.FragmentActivity;
+
+import androidx.fragment.app.FragmentActivity;
 
 public class BaseFragmentActivity extends FragmentActivity {
 
